Add Project README
Add a README file that identifies the purpose of the project alongside
various badges, one of which provides realtime feedback on the status
of the CI builds run via Travis-CI and another which provides a quick
and friendly indicator that the project is MIT licensed and a linked
copy of said license. Additionally the README clearly displays another
copyright notice for Emily Mabrey for the year 2016. The README is
written in Markdown which allows plain text comprehension while also
allowing GitHub to parse the README for use as the project's repository
description.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>