Create emabrey-parent-resources module
Add the emabrey-parent-resources project module with multiple build
resources that will be needed by the emabrey-parent POM build. These
resources include: a customized CSS file resource for use in the
customization of the JavaDoc reports generated by emabrey-parent
child projects; a template of the Emabrey Proprietary License (EPL),
which will be used to automatically verify and format the license and
copyright headers of various child project files. Automating the
header formatting makes it easy to avoid header mismatches and to
automate copyright changes for long-lived projects. The main benefit
to using this module to distribute the build resources is the builds
remain deterministic and are much more portable and user-friendly.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>