Add Travis-CI configuration which uses Docker
Add a Travis-CI configuration file which utilizes a custom Docker image
containing a specified version of Maven and a specified JDK version
to build and test the project in a sterile Docker container. The
Docker container currently contains both Maven 3.3.9 and OpenJDK 8
via the executed Dockerfile. The Dockerfile generates an image with
the name of "emabrey-parent/emabrey-default-maven-image" upon the
execution of each Travis-CI build. For maximum performance the Maven
M2 repository is cached between builds and provided to the running
Docker image. This should dramatically reduce build times, but it
might contaminate the otherwise sterile Docker containers. Since Maven
verifies artifact hashes though, that risk is slim. The Dockerfile
relies on the openJDK Docker registry/image and on the availability of
certain files from the carlossg/docker-maven GitHub repository.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>