Add MAINTAINER metadata to Dockerfile
Simply add the MAINTAINER metadata to the emabrey-default-maven-image
image's Dockerfile. The MAINTAINER metadata functions much like an
author section for that particular Dockerfile, and as such, it has
been configured to show that Emily Mabrey is the maintainer of the
Dockerfile. The email address for Emily Mabrey is also included per the
standard format of the MAINTAINER field.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>