Fix Travis-CI configuration
Remove multiple syntax errors and increase the readability of the
Travis-CI config file ".travis.yml". Changes include the restructure
of the build lifecycle to match the recommended two-step build of
install followed by script, the fix of a typo in the defining of certain
environmental variables, the removal of some environmental variables
that were verbose and also the reorganization of the text to be more
concise and more clearly delineated. This commit will hopefully fix
the errors which caused the first attempt at a Travis-CI build to fail,
as the next Travis-CI build should not attempt to execute the
environmental variable declarations as though they were commands as the
previous one unexpectedly did.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>