Attempt to fix Travis-CI configuration again
Remove the environmental variables causing build errors completely.
Rename the Docker build runtime name to clarify the run is executing
the Travis-CI build. Alter the declared language to be Java to more
clearly reflect that the build is in fact a Java/Maven project, despite
the fact that Docker is used heavily during the Travis-CI build process.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>