Fix Dockerfile by making a script executable
Utilizing a RUN call to the chmod program the downloaded bash script is
now explicitly marked as being executable. Previously the file was
downloaded and then run without verifying the metadata as it was given
originally (from the downloaded file data) was valid. This RUN fixes the
error that caused the previous Travis-CI build error. Though there was
no error for this particular executable before, the downloaded Maven
binary is also being explicitly marked as an executable file. Additional
but minor housekeeping changes were performed in an effort to reorganize
the code so that it clearly separates the commands run upon one file
from the commands run upon another file.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>