Fix Dockerfile to correctly install Maven
Previously the code was written assuming that the ADD command both
downloaded files from a remote or local source and also uncompressed
files from a remote or local source. The actual functionality is that
files can be added from remote or local sources but file decompression
is only available for locally sourced files provided to the ADD
command. As a result a manual RUN call to untar and ungzip the already
download maven binary tarball was added to the Dockerfile. Also, to
clear up the appearance of the Dockerfile, the longest variable name was
shortened. The only change to the Travis-CI configuration was the change
of the language from Java to Bash. Reverting the language to Bash should
avoid the long loading of the Java language build environment. The
specification of the Bash language also ensures the Dockerfile has the
required access to a Bash compatible shell. Because the Java and Maven
tools provided via the build environment are being provided via Docker,
this language change should have no effect on the build other than to
shorten it somewhat.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>