Fix Dockerfile RUN arguments given to tar
The previous commit correctly instructed the Dockerfile to run the tar
program with the xzC options, indicating that it should decompress and
untar the given tarball while stripping a given number of components.
Unfortunately, the options lacked the f option, which indicates the
tarball is actually a file which should be loaded from the disk. Because
that f option was absent the tar program attempted to load the tarball
from STDIN, and that of course failed. This commit adds the f option and
corrects the tar program execution.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>