Fix Dockerfile RUN tar command again
The previous attempt at fixing the RUN of the tar program with the
added f option failed with the same "unexpected end of file on STDIN"
error that it was intended to fix. An alternate attempt to fix the
execution of the tar program is the addition of the cat program. The
cat program reads a file and outputs it's contents to STDIN, the same
place that the tar program is complaining it cannot find the data it
needs. By running cat to output the tarball to the console and then
piping that tarball console data to the tar command the problem causing
the error will hopefully be resolved.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>