Fix and upgrade Dockerfile
Fix the RUN commands by changing them to the exec format, which silently
passes the commands to the configured shell. The exec format allows
multiline declarations, which also brings the benefit of faster builds
(by avoiding making more intermediate images), so all the RUN commands
have been converted to be declared using multiple lines. Comments which
explain the function of specific declarations have been added.
Additionally, by leveraging the inner functionality of the ARG and ENV
commands, the build time variables are now user configurable with the
build-args option of "docker run", but they are also constant values
during the actual script execution. This should help prevent confusion
over the somewhat hard to understand scoping of ARG declarations. Those
constant values are set to sensible defaults should the user not specify
any build-args, so the user only needs to use the build-arg option if the
defaults are not acceptable. The Maven configuration transfer from host
to the image was made obvious by the inclusion of a VOLUME declaration
which explicitly indicates the user has an available mount point for the
Maven configuration directory of the host machine.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>