Revert to standard Travis-CI configuration
After many attempts to get the Docker based configuration working via
Travis-CI, this commit removes the Dockerfile and changes the Travis-CI
configuration to use a more standard style build. The new Travis-CI
build configuration includes support for using the latest versions of
openJDK8 and oracleJDK8 independent of the versions normally available
on Travis-CI build machines. This is accomplished by manual apt-get
calls to install the most recently available versions of those JDKs,
though it should be noted that there is a significant lag between
a new JDK version releasing and a JDK version being back-ported to the
version of Ubuntu (trusty) that the build machines use.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>