Improve and refine Travis-CI configuration
Instead of using the env key as an improper stand-in for the built in
matrix feature, replace the env key values with an equivalent matrix
declaration. Modify the execution of the jdkswitcher program to only
occur if the custom_jdk variable is true. Controlling the execution of
jdk_switcher with a conditional should hopefully allow use to use the
built in jdk key for non-custom builds with officially supported JDK
versions and vendors. Before this change the jdkswitcher would smudge
the build machines JAVA_HOME in a way that might break official
Travis-CI Java builds.

A fix to the apt-get update command was made by adding a -y option;
previously the command would have waited for the interactive input of
a yes/no question, which would have broken the build (since the
Travis-CI builder is not interactive). All builds now explicitly state
that they should be built via the Linux OS with root access available
via sudo. An attempt was made to remove the need for sudo use and switch
to the much container builders (which have much faster spinup), but the
apt addon (the only alternative to using sudo) does not support the
needed Bash shell conditionals we use to allow both officially supported
and unsupported JDK versions.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>