Fix Travis-CI build matrix
By using the language:generic configuration and allowing the built-in
and officially supported JDKs to run without any custom changes, this
modified configuration should not have any errors related to having
unknown build types, etc. . Additionally, the structure of the config
was changed to be more readable: all apt-get preparation is done in
before_install and all apt-get installation is now done in install.
Before those two build sections we not well distinguished, with some
preparation for install actually occurring in the install section
instead of the before_install section.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>