Move plugin executions in emabrey-parent POM
Reorganize the POM of the emabrey-parent project to separate the
configurations from the executions. The parent POM is intended to pass
down default configurations but it is outside the scope of the parent
POM to setup executions which children will inherit, especially since
the children may have a slightly different lifecycle and other reasons
to not execute the plugins exactly the way the default does. Instead of
requiring children to memorize which executions to override and what the
execution ids are etc., it is far simpler and better to leave it up to
child POMs what to execute and when to execute it. As a result of
removing the execution configurations from the pluginManagement section
the plugins executed during emabrey-parent's self-build have been moved
to the project specific plugins section. The details of what executes
and when remains identical to the previously inheritable defaults, but
they are simply not inheritable now. Modify the emabrey-parent README to
have improved alternate text for the info badges if they fail to render
or for users viewing the project README with, for example, a
screen-reader.

Add several plugin executions to the emabrey-parent-resources POM,
including the license plugin execution modified to use the MIT license
header, the GPG plugin to produce signed artifacts, and the JavaDoc and
Source plugins to generate secondary artifacts that make the
emabrey-project-resources project compatible for binary distribution
via Maven Central etc.. Also in the emabrey-project-resources project,
add an MIT license header to the license resources and fix a minor
text misalignment issue in the javadoc.css file.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>