Reattempt fix of installAptPrograms script
Correcting the typo in the use of jdk_switcher still failed, so the
entire call to jdk_switcher has been removed. It is not clear if the
build will succeed or fail without using jdk_switcher, but hopefully
it will succeed since we are manually installing only one JDK to the
build system.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>