Modify the GPG and compiler plugins
Fixed the execution configuration of the GPG plugin in both POMs by
removing the configured executable value completely. The plugin will
now just default to the standard executable instead of looking for
one by the specified name.

Also, modified the compiler plugin by updating the version from 3.5.1
to 3.6.0 and targeting the javac compiler with the errorprone static
analysis add-in via the compilerId of javac-with-errorprone. The
compiler plugins were modified within both POMs.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>