Fix Travis-CI artifact deployment
Correct the incorrect Travis-CI deployment to correctly deploy for the
develop and master branches instead of exclusively for the master
branch. The develop branch should generate build deployments so that
we have a record of the build JARs of specific commits within
Artifactory. We can then manually choose which versions should be
publicly distributed via exporting to Bintray.

Also, remove the git-flow Maven plugin because given the changes to the
deployment process it is no longer relevant to the release process of
this project.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>