Reattempt fix of Artifactory deployment setup
Add the artifactory-maven-plugin to the emabrey-parent-resources POM
using the same configuration used in the main emabrey-parent POM. The
properties used in that plugins configuration were also transplanted
into the emabrey-parent-resources POM. In this case the
artifactory-maven-plugin was manually added, but the plugins declaration
in emabrey-parent has been modified to automatically inherit for the
children of the parent POM.

Additionally, minor modifications were made to the maven-javadoc-plugin
by redefining the bottom text of generated reports to contain a CDATA
section that includes a link to the "emabrey" GitHub profile. Using the
CDATA section also allowed the '©' character to be used in the literal
form instead of as a Unicode escape sequence.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>