Add enforcer and resources plugins to both POMs
Add the maven-enforcer and the maven-resources plugins to the POMs of
the emabrey-parent and emabrey-parent-resources projects. The
maven-enforcer plugin uses custom rules provided by the
extra-enforcer-rules plugin dependency. The maven-resources-plugin
ensures the project license and README files are correctly included
within the built project artifacts, which should allow Artifactory to
auto-detect license compatibility for our uploads.

The emabrey-parent POM was reorganized and had several minor typos fixed
to improve it's readability. Typos fixed include: the
license-maven-plugin incorrectly using a groupId property as a version,
making the maven-javadoc-plugin default to the maven stylesheet as was
originally intended, and removal of the outdated comments for the
maven-gpg-plugin. Additionally properties containing the Maven and Java
versions enforced/required by the POM. A Minor fix also was made in the
Travis-CI deploy script by removing outdated comments.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>