Fix usage of "update-java-alternatives" script
Fix the incorrect usage of the "update-java-alternatives" script by
replacing the usage of the custom JDK directory with the custom JDK
identifier. Additionally, add apt-get commands to remove the default
JDK installs from any build which we are installing a custom JDK into-
this should help prevent misuse of the default JDKs in the future, as
they cannot be used if they are removed entirely from the build
environment. Finally, on advice of the apt-get MAN page, all usages of
the "-qq" option have been reduced to a single "-q", as "-qq" makes the
execution completely silent, which is not desirable for fixing build
errors or reproducing failures.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>