Fix JAVA_HOME Maven error and OracleJDK8 install
Use the "debconf-set-selections" command to accept the Oracle license
that was causing the OracleJDK8 install to fail because we had not
accepted/read the license, which by default requires interaction which
we could not provide. Additionally, using a clunky hack involving a
symlink, point the default JAVA_HOME directory to our custom install, so
all pre-configured Travis-CI usages of JAVA_HOME will now point to our
custom "JAVA_HOME".

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>