Fix and cleanup Travis-CI build config
Removed the Java 7 execution from .travis.yml because it was not
compatible with the enforcement rules of the project POMs and modified
the variables used to control the custom JDK install:

OVERRIDE_JDK => INSTALL_JDK
JDK => [removed]

Cleaned and refactored the Travis-CI build scripts, including
addAptRepos and installAptPrograms, and added various sanity checks and
input verifications. Now if a validation error is encountered an error
message is printed to the output log of the CI build. Improved support
for custom JDK installations, including support for Java 6, Java 7,
Java 8, and the as yet unreleased Java 9 JDK.

Made an additional minor change to the emabrey-parent POM- the
"versions.java.minimum" property was renamed to
"versions.buildJDK.minimum". This modification will make the distinction
between the build JDK and the target JDK more obvious should future
children wish to target a different version via backwards compatibility.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>