Fix emabrey-parent-resources POM
The emabrey-parent-resources was missing the repository and
pluginRepository declarations needed to allow Maven to find and use the
JCenter Maven repository. Those repository declarations have been added
to the POM, which should correct the build problems caused by Maven
being unable to find the artifactory-maven-plugin in the standard Maven
central repository.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>