Fix installMavenWrapper script
The "mvnw" command is not added to the PATH, so we now call it
explicitly via the new Bash variable ${MVN}, which is exported by the
installMavenWrapper script. An additional ${MVN_NAME} variable is also
defined for logging purposes. The scripts have been modified to make
explicit calls to "mvnw" if it is successfully installed, or otherwise
to fall-back onto the default usage of the "mvn" command if the ${MVN}
variable is unset. It is possible the fall-back usage of "mvn" may fail
during execution of the Maven POM, but it may succeed sometimes as well.
Rather then failing pessimistically we continue the execution hoping the
pre-installed Maven version available via the Travis-CI container will
suffice.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>