Reattempt adding mvnw command support
Remove the changes to the PATH and instead replace them with a manual
variable addition to the .travis.yml file. The new variable, MVNW_CMD,
simply points directly to the command files location using an absolute
path. If the command is not successfully installed or successfully
executed the support for falling back to the default Travis-CI Maven
is still included.

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>